// src/generated/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const BankGate = __t.object("BankGate", {
  owner: __t.identity(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  color: __t.string(),
  sealedUntilTick: __t.i64(),
});
export type BankGate = __Infer<typeof BankGate>;

export const Game = __t.object("Game", {
  id: __t.u32(),
  phase: __t.string(),
  nowTick: __t.i64(),
  endsTick: __t.i64(),
  winner: __t.option(__t.identity()),
  relicsTotal: __t.u32(),
});
export type Game = __Infer<typeof Game>;

export const GameTickSchedule = __t.object("GameTickSchedule", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type GameTickSchedule = __Infer<typeof GameTickSchedule>;

export const InputState = __t.object("InputState", {
  forward: __t.bool(),
  backward: __t.bool(),
  left: __t.bool(),
  right: __t.bool(),
  sprint: __t.bool(),
  jump: __t.bool(),
  attack: __t.bool(),
  castSpell: __t.bool(),
  sequence: __t.u32(),
});
export type InputState = __Infer<typeof InputState>;

export const Keeper = __t.object("Keeper", {
  id: __t.u32(),
  lastLine: __t.string(),
  lastAction: __t.string(),
  lastTarget: __t.string(),
  seq: __t.u32(),
  updatedTick: __t.i64(),
});
export type Keeper = __Infer<typeof Keeper>;

export const LoggedOutPlayerData = __t.object("LoggedOutPlayerData", {
  identity: __t.identity(),
  username: __t.string(),
  characterClass: __t.string(),
  color: __t.string(),
  score: __t.i32(),
  lastSeen: __t.timestamp(),
});
export type LoggedOutPlayerData = __Infer<typeof LoggedOutPlayerData>;

export const PlayerData = __t.object("PlayerData", {
  identity: __t.identity(),
  username: __t.string(),
  characterClass: __t.string(),
  get position() {
    return Vector3;
  },
  get rotation() {
    return Vector3;
  },
  health: __t.i32(),
  maxHealth: __t.i32(),
  mana: __t.i32(),
  maxMana: __t.i32(),
  currentAnimation: __t.string(),
  isMoving: __t.bool(),
  isRunning: __t.bool(),
  isAttacking: __t.bool(),
  isCasting: __t.bool(),
  lastInputSeq: __t.u32(),
  get input() {
    return InputState;
  },
  color: __t.string(),
  carryingRelicId: __t.option(__t.u64()),
  score: __t.i32(),
  speedMult: __t.f32(),
  cursedUntilTick: __t.i64(),
  blessedUntilTick: __t.i64(),
  blessMult: __t.f32(),
  stunnedUntilTick: __t.i64(),
  lastCurseDropTick: __t.i64(),
});
export type PlayerData = __Infer<typeof PlayerData>;

export const Relic = __t.object("Relic", {
  id: __t.u64(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  state: __t.string(),
  carriedBy: __t.option(__t.identity()),
});
export type Relic = __Infer<typeof Relic>;

export const Trap = __t.object("Trap", {
  id: __t.u64(),
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
  armed: __t.bool(),
  createdTick: __t.i64(),
});
export type Trap = __Infer<typeof Trap>;

export const Vector3 = __t.object("Vector3", {
  x: __t.f32(),
  y: __t.f32(),
  z: __t.f32(),
});
export type Vector3 = __Infer<typeof Vector3>;